Rename/reorganize components folder.
