+++ src/components/Artist/Card.tsx
