+++ src/components/Artist/Modal.tsx
