--- src/components/artist/index.ts
@@ -1,3 +1,0 @@
-import ArtistGallery from "./ArtistGallery";
-
-export { ArtistGallery };
