--- src/components/musicPlayer/index.ts
@@ -1,3 +1,0 @@
-import { useMusicPlayer, MusicPlayerProvider } from './MusicPlayerContext';
-
-export { useMusicPlayer, MusicPlayerProvider };
