--- src/components/track/index.ts
@@ -1,3 +1,0 @@
-import TrackGallery from "./TrackGallery";
-
-export { TrackGallery };
